Treat measure width as double
Measure width attribute can be a decimal value, and does not have to be an integer.  This was discovered when parsing a music xml file generated by MuseScore2.

+++ MusicXml/Domain/Measure.cs
@@ -10,7 +10,7 @@ namespace MusicXml.Domain
 			MeasureElements = new List<MeasureElement>();
 		}
 
-		public int Width { get; set; }
+		public double Width { get; set; }
 		
 		// This can be any musicXML element in the measure tag, i.e. note, backup, etc
 		public List<MeasureElement> MeasureElements { get; private set; }
